django-api/requirements.txt
cryptography
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
hedera-sdk-py
PyJWT==2.1.0
tinyec